Calculate and add the transition disruption metric to a dataset

This function takes a typical PACTA output, and uses it to calculate the transition disruption metric.

Usage

calculate_tdm(
  data,
  t0,
  delta_t1 = 5,
  delta_t2 = 10,
  additional_groups = NULL,
  scenarios = "IPR FPS 2021",
  log_path = .GlobalEnv$log_path
)

Arguments

data

A data.frame (or tibble), with a format similar to the *_results_portfolio.rds outputs of PACTA.

t0

The start year, from which the TDM value should be calculated. This value should coincide with the earliest year in the PACTA results dataset.

delta_t1

The number of years into the future for which there is production data. The default value is 5 years since this is often the limit of the forward-looking data.

delta_t2

The number of years into the future against which you want to calculate disruption.

additional_groups

Character vector. The names of columns to group by, in addition to these ones (which are always used):

technology, ald_sector

scenarios

Character vector. The names of scenarios for which TDM should be calculated.

log_path

String. A path leading to the desired logfile output.

Value

A tibble with the columns specified in the additional_groups input as well as tdm_technology_value: the technology level transition disruption metric, tdm_sector_value: the sector level transition disruption metric, tdm_portfolio_value: the portfolio level transition disruption metric, tdm_metric: a flag that indicates what each tdm_*_value is being calculated against (e.g. "portfolio" or "scenario"), as well as tdm_t0, tdm_delta_t1 and tdm_delta_t2: corresponding to the input arguments.

Examples

library(tibble)

pacta_results <- tibble(
  scenario = rep("IPR FPS 2021", 3),
  allocation = rep("portfolio_weight", 3),
  equity_market = rep("GlobalMarket", 3),
  scenario_geography = rep("Global", 3),
  ald_sector = rep("Power", 3),
  technology = rep("RenewablesCap", 3),
  year = c(2020, 2025, 2030),
  plan_alloc_wt_tech_prod = rep(1, 3),
  scen_alloc_wt_tech_prod = 1:3,
  plan_carsten = rep(0.5, 3)
)
pacta_results
#> # A tibble: 3 × 10
#>   scenario     allocation equity_market scenario_geography ald_sector technology
#>   <chr>        <chr>      <chr>         <chr>              <chr>      <chr>     
#> 1 IPR FPS 2021 portfolio… GlobalMarket  Global             Power      Renewable…
#> 2 IPR FPS 2021 portfolio… GlobalMarket  Global             Power      Renewable…
#> 3 IPR FPS 2021 portfolio… GlobalMarket  Global             Power      Renewable…
#> # ℹ 4 more variables: year <dbl>, plan_alloc_wt_tech_prod <dbl>,
#> #   scen_alloc_wt_tech_prod <int>, plan_carsten <dbl>

calculate_tdm(pacta_results, t0 = 2020)
#> # A tibble: 2 × 9
#>   technology    ald_sector tdm_metric tdm_technology_value tdm_sector_value
#>   <chr>         <chr>      <chr>                     <dbl>            <dbl>
#> 1 RenewablesCap Power      portfolio                     2                2
#> 2 RenewablesCap Power      scenario                      1                1
#> # ℹ 4 more variables: tdm_portfolio_value <dbl>, tdm_t0 <dbl>,
#> #   tdm_delta_t1 <dbl>, tdm_delta_t2 <dbl>